+++ src/apis/price-notification/entitites/price-notification.entity.ts
