+++ src/app.module.ts
@@ -17,7 +17,7 @@ import { APP_FILTER, APP_INTERCEPTOR, APP_PIPE } from '@nestjs/core';
 import { TokenModule } from './token/token.module';
 import { DateUtilModule } from './utils/date-util/date-util.module';
 import { ExchangeRateModule } from './apis/exchange-rate/exchange-rate.module';
-import { PriceNotificationModule } from './apis/price-alert/price-notification.module';
+import { PriceNotificationModule } from './apis/price-notification/price-notification.module';
 import { AlertHistoryModule } from './apis/alert-history/alert-history.module';
 import { NewsModule } from './apis/news/news.module';
 import { WatchlistModule } from './apis/watchlist/watchlist.module';
